Split Lists.module.scss into parts by component
