+++ packages/slate-editor/src/modules/editor-v4-rich-formatting/components/Lists/ListItemElement.module.scss
@@ -1,0 +1,5 @@
+@import "styles/variables";
+
+.ListItemElement {
+    margin: $spacing-1 0;
+}
