+++ packages/slate-editor/src/modules/editor-v4-rich-formatting/components/Lists/ListItemTextElement.module.scss
@@ -1,0 +1,8 @@
+@import "styles/variables";
+
+.ListItemTextElement {
+    color: $editor-v4-text-color;
+    font-weight: 500;
+    font-size: $editor-v4-paragraph-font-size;
+    line-height: $editor-v4-paragraph-line-height;
+}
